Wave timer is now a float rather than an int. Update my scene.

+++ Button Bash/Assets/Scripts/WaveInformation.cs
@@ -4,9 +4,13 @@ using UnityEngine;
 
 public class WaveInformation : MonoBehaviour
 {
-	// The enemies in the wave.
+	/// <summary>
+	/// The enemies in the wave.
+	/// </summary>
 	public GameObject[] m_WaveEnemies;
 
-	// The timers to spawn the enemies.
-	public int[] m_WaveEnemySpawnTimes;
+	/// <summary>
+	/// The timers to spawn the enemies.
+	/// </summary>
+	public float[] m_WaveEnemySpawnTimes;
 }
